pyproject.toml
[project]
name = "chatten"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "renardeinside", email = "polarpersonal@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "dash>=2.18.2",
    "databricks-langchain>=0.2.0",
    "databricks-sdk>=0.43.0",
    "databricks-sql-connector>=4.0.0",
    "fastapi>=0.115.8",
    "langchain>=0.3.17",
    "langchain-community>=0.3.16",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "wheel>=0.45.1",
]
